requirements.txt
telethon
cryptg
python-gnupg
